Console Applications

+++ ConsoleApplications/ReservationSystem/BookingServices.java
@@ -3,7 +3,7 @@ package ConsoleApplications.ReservationSystem;
 public interface BookingServices {
     abstract void bookTicket(String name, int age);
 
-    abstract void cancelTicket(int id);
+    abstract void cancelTicket(long id);
 
-    abstract void printTicket(int id);
+    abstract void printTicket(long id);
 }
